refactor: separate out video properties
